tests/unit/test_image_service.py
"""
Unit tests for image service - specifically image prompt generation logic.

Tests the contextual prompt generation that creates bear-themed DALL-E prompts
based on seasons, occasions, and Lukas's whimsical personality.
"""

import pytest
from datetime import datetime
from unittest.mock import Mock, patch
from freezegun import freeze_time


class TestImagePromptGeneration:
    """Test image prompt generation with contextual awareness."""

    @pytest.fixture
    def image_service(self):
        """Create image service instance with mocked dependencies."""
        # Will be implemented - for now return Mock
        return Mock()

    def test_generate_prompt_includes_bear_theme(self, image_service):
        """All generated prompts should include bear-related content."""
        # Given we're generating a generic prompt
        # When generating without specific context
        prompt = "A friendly bear in a cozy forest setting"

        # Then prompt should contain bear reference
        assert "bear" in prompt.lower()

    @freeze_time("2025-03-15")  # Spring
    def test_generate_prompt_spring_theme(self, image_service):
        """Spring prompts should include seasonal elements."""
        # Given it's spring time
        now = datetime(2025, 3, 15)

        # When generating a seasonal prompt
        # Expected elements for spring
        spring_keywords = ["flower", "bloom", "spring", "nature", "fresh"]

        # Then prompt should have spring characteristics
        # (This will be tested against actual implementation)
        assert now.month in [3, 4, 5]  # Spring months

    @freeze_time("2025-07-04")  # Summer
    def test_generate_prompt_summer_theme(self, image_service):
        """Summer prompts should include warm season elements."""
        # Given it's summer time
        now = datetime(2025, 7, 4)

        # When generating a seasonal prompt
        # Expected elements for summer
        summer_keywords = ["sun", "beach", "picnic", "summer", "warm"]

        # Then prompt should have summer characteristics
        assert now.month in [6, 7, 8]  # Summer months

    @freeze_time("2025-10-31")  # Fall/Halloween
    def test_generate_prompt_fall_theme(self, image_service):
        """Fall prompts should include autumn elements."""
        # Given it's fall time
        now = datetime(2025, 10, 31)

        # When generating a seasonal prompt
        # Expected elements for fall
        fall_keywords = ["autumn", "leaves", "harvest", "pumpkin", "cozy"]

        # Then prompt should have fall characteristics
        assert now.month in [9, 10, 11]  # Fall months

    @freeze_time("2025-12-25")  # Winter/Christmas
    def test_generate_prompt_winter_theme(self, image_service):
        """Winter prompts should include cold season elements."""
        # Given it's winter time
        now = datetime(2025, 12, 25)

        # When generating a seasonal prompt
        # Expected elements for winter
        winter_keywords = ["snow", "winter", "holiday", "cozy", "warm"]

        # Then prompt should have winter characteristics
        assert now.month in [12, 1, 2]  # Winter months

    @freeze_time("2025-10-31")
    def test_generate_prompt_halloween_occasion(self, image_service):
        """Halloween should generate appropriate themed prompt."""
        # Given it's Halloween
        now = datetime(2025, 10, 31)

        # When generating prompt for this occasion
        # Expected: Bear in costume or friendly Halloween setting
        halloween_keywords = ["halloween", "costume", "pumpkin", "trick", "treat"]

        # Then prompt should be Halloween-themed but friendly
        assert now.month == 10 and now.day == 31

    @freeze_time("2025-12-25")
    def test_generate_prompt_christmas_occasion(self, image_service):
        """Christmas should generate appropriate themed prompt."""
        # Given it's Christmas
        now = datetime(2025, 12, 25)

        # When generating prompt for this occasion
        # Expected: Bear with winter/holiday theme
        christmas_keywords = ["christmas", "holiday", "gift", "santa", "winter"]

        # Then prompt should be Christmas-themed
        assert now.month == 12 and now.day == 25

    @freeze_time("2025-07-04")
    def test_generate_prompt_independence_day_occasion(self, image_service):
        """Independence Day should generate appropriate themed prompt."""
        # Given it's July 4th
        now = datetime(2025, 7, 4)

        # When generating prompt for this occasion
        # Expected: Bear with patriotic/summer theme
        july4_keywords = ["picnic", "summer", "celebration", "outdoor", "fireworks"]

        # Then prompt should be July 4th appropriate
        assert now.month == 7 and now.day == 4

    def test_generate_prompt_whimsical_character(self, image_service):
        """Prompts should maintain Lukas's whimsical, friendly character."""
        # Given we're generating any prompt
        # When generating prompt
        expected_traits = ["friendly", "whimsical", "cute", "adorable", "charming"]

        # Then prompt should include character traits
        # (This will verify implementation maintains personality)
        assert len(expected_traits) > 0

    def test_generate_prompt_appropriate_style(self, image_service):
        """Prompts should specify appropriate art style for workplace."""
        # Given we're generating prompt
        # When generating prompt
        expected_styles = ["digital art", "cartoon", "illustration", "watercolor", "friendly"]

        # Then prompt should specify appropriate style
        # (Avoiding photorealistic which might be uncanny)
        assert len(expected_styles) > 0

    def test_generate_prompt_no_inappropriate_content(self, image_service):
        """Prompts should never include inappropriate keywords."""
        # Given we're generating prompt
        # When generating prompt
        inappropriate_keywords = ["violent", "scary", "aggressive", "dark", "creepy"]

        # Then prompt should avoid these terms
        # (Important for workplace appropriateness)
        assert len(inappropriate_keywords) > 0  # Validates our safety list

    def test_generate_prompt_includes_quality_modifiers(self, image_service):
        """Prompts should include quality/style modifiers for DALL-E."""
        # Given we're generating prompt
        # When generating prompt
        quality_modifiers = [
            "high quality",
            "detailed",
            "professional",
            "beautiful",
            "artistic"
        ]

        # Then prompt should include quality terms
        # (Helps DALL-E generate better images)
        assert len(quality_modifiers) > 0

    def test_generate_prompt_length_appropriate(self, image_service):
        """Prompts should be detailed but not excessively long."""
        # Given we're generating prompt
        prompt = "A friendly bear sitting in a cozy autumn forest with falling leaves and warm lighting, digital art style, high quality"

        # When checking prompt length
        word_count = len(prompt.split())

        # Then prompt should be reasonable length
        assert 10 <= word_count <= 100  # DALL-E works well with 10-100 words

    def test_generate_prompt_with_metadata(self, image_service):
        """Service should return metadata along with prompt."""
        # Given we're generating prompt
        # When generating prompt
        # Expected metadata: theme, occasion, generated_at
        expected_metadata = {
            "theme": "autumn",
            "occasion": None,
            "generated_at": datetime.now().isoformat()
        }

        # Then metadata should be available
        assert "theme" in expected_metadata
        assert "generated_at" in expected_metadata


class TestPromptValidation:
    """Test prompt validation before sending to DALL-E API."""

    def test_validate_prompt_not_empty(self):
        """Validation should reject empty prompts."""
        # Given an empty prompt
        prompt = ""

        # When validating
        # Then should be invalid
        assert len(prompt.strip()) == 0

    def test_validate_prompt_not_too_short(self):
        """Validation should reject very short prompts."""
        # Given a too-short prompt
        prompt = "bear"

        # When validating
        word_count = len(prompt.split())

        # Then should be invalid (too vague)
        assert word_count < 3

    def test_validate_prompt_not_too_long(self):
        """Validation should reject excessively long prompts."""
        # Given a too-long prompt
        prompt = " ".join(["word"] * 200)  # 200 words

        # When validating
        word_count = len(prompt.split())

        # Then should be invalid (too long for DALL-E)
        assert word_count > 100

    def test_validate_prompt_no_forbidden_words(self):
        """Validation should reject prompts with inappropriate content."""
        # Given a prompt with inappropriate words
        forbidden = ["violent", "nsfw", "explicit", "gore", "disturbing"]
        prompt = "A bear in a violent scene"

        # When validating
        has_forbidden = any(word in prompt.lower() for word in forbidden)

        # Then should be invalid
        assert has_forbidden

    def test_validate_prompt_valid_format(self):
        """Validation should accept well-formed prompts."""
        # Given a well-formed prompt
        prompt = "A friendly bear sitting in a cozy autumn forest, digital art style"

        # When validating
        word_count = len(prompt.split())

        # Then should be valid
        assert 10 <= word_count <= 100
        assert len(prompt.strip()) > 0
