pyproject.toml
[project]
name = "lukas-bear-chatbot"
version = "0.1.0"
description = "Lukas the Bear - A friendly Slack chatbot mascot powered by AI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Team", email = "team@example.com" }
]

dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.23.0",
    "aiohttp>=3.9.0",  # Required for AsyncApp
    "openai>=1.3.0",
    "apscheduler>=3.10.4",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "tenacity>=8.2.3",
    "pybreaker>=1.0.1",
    "tiktoken>=0.5.1",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    # MCP (Model Context Protocol) integration
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]  # Line too long - we use 100 but allow some flexibility
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
